// admin-frontend/src/api/generated/model/systemInfoVO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin Backend API
 * 后台管理系统后端 API 文档，前端可通过 OpenAPI Schema 自动生成接口调用代码
 * OpenAPI spec version: 1.0.0
 */
import type { SystemCpuInfo } from './systemCpuInfo';
import type { SystemDbPoolInfo } from './systemDbPoolInfo';
import type { SystemDiskInfo } from './systemDiskInfo';
import type { SystemJvmInfo } from './systemJvmInfo';
import type { SystemMemoryInfo } from './systemMemoryInfo';
import type { SystemRedisInfo } from './systemRedisInfo';
import type { SystemServerInfo } from './systemServerInfo';

/**
 * 系统监控信息
 */
export interface SystemInfoVO {
  cpu?: SystemCpuInfo;
  memory?: SystemMemoryInfo;
  jvm?: SystemJvmInfo;
  /** 磁盘信息 */
  disks?: SystemDiskInfo[];
  redis?: SystemRedisInfo;
  dbPool?: SystemDbPoolInfo;
  server?: SystemServerInfo;
}